+++ internal/cmd/write.go
@@ -1,10 +1,14 @@
 package cmd
 
 import (
+	"encoding/json"
 	"fmt"
 	"io"
+	"log"
 	"net/http"
 	"os"
+
+	"github.com/dhuan/mock/internal/utils"
 
 	"github.com/spf13/cobra"
 )
@@ -19,6 +23,27 @@ var writeCmd = &cobra.Command{
 			}
 
 			contentToWrite := stdin
+
+			if flagJson {
+				if flagAppend {
+					fmt.Printf("--json cannot be used with --append.\n")
+
+					os.Exit(1)
+				}
+
+				jsonStr, err := formatJson(contentToWrite)
+				if err != nil || string(jsonStr) == "null" {
+					os.Exit(1)
+				}
+
+				if err = utils.AddLineToFile(rf.headers, "Content-Type: application/json"); err != nil {
+					log.Print(err)
+
+					os.Exit(1)
+				}
+
+				contentToWrite = jsonStr
+			}
 
 			if flagAppend {
 				fileContent, err := os.ReadFile(os.Getenv("MOCK_RESPONSE_BODY"))
@@ -37,4 +62,11 @@ var writeCmd = &cobra.Command{
 			}
 		})
 	},
+}
+
+func formatJson(data []byte) ([]byte, error) {
+	var result interface{}
+	json.Unmarshal(data, &result)
+
+	return json.Marshal(result)
 }