+++ tests/e2e/write_test.go
@@ -61,3 +61,71 @@ func Test_E2E_Write_WritingMultipleTimesOverwrites(t *testing.T) {
 		StringMatches("Write again."),
 	)
 }
+
+func Test_E2E_Write_WithJsonOption_Ok(t *testing.T) {
+	RunTestWithNoConfigAndWithArgs(
+		t,
+		[]string{
+			"--route foo/bar",
+			fmt.Sprintf("--exec '%s'", strings.Join([]string{
+				// Space excess is intentional in order to assert
+				// that --json formats the JSON for us.
+				`printf "{\"foo\":          \"bar\"}" | {{MOCK_EXECUTABLE}} write --json`,
+				`mock set-header Exit-Status-Code "${?}"`,
+			}, ";")),
+		},
+		"GET",
+		"foo/bar",
+		nil,
+		nil,
+		StringMatches(`{"foo":"bar"}`),
+		HeadersMatch(map[string][]string{
+			"Content-Type":     {"application/json"},
+			"Exit-Status-Code": {"0"},
+		}),
+	)
+}
+
+func Test_E2E_Write_WithJsonOption_InvalidJson(t *testing.T) {
+	RunTestWithNoConfigAndWithArgs(
+		t,
+		[]string{
+			"--route foo/bar",
+			fmt.Sprintf("--exec '%s'", strings.Join([]string{
+				`printf "{\"foo\":INVALID}" | {{MOCK_EXECUTABLE}} write --json`,
+				`mock set-header Exit-Status-Code "${?}"`,
+			}, ";")),
+		},
+		"GET",
+		"foo/bar",
+		nil,
+		nil,
+		StringMatches(``),
+		HeadersMatch(map[string][]string{
+			"Exit-Status-Code": {"1"},
+		}),
+		HeaderKeysNotIncluded([]string{"Content-Type"}),
+	)
+}
+
+func Test_E2E_Write_WithJsonOption_CannotUseWithAppend(t *testing.T) {
+	RunTestWithNoConfigAndWithArgs(
+		t,
+		[]string{
+			"--route foo/bar",
+			fmt.Sprintf("--exec '%s'", strings.Join([]string{
+				`printf "{\"foo\":123}" | {{MOCK_EXECUTABLE}} write --json --append`,
+				`mock set-header Exit-Status-Code "${?}"`,
+			}, ";")),
+		},
+		"GET",
+		"foo/bar",
+		nil,
+		nil,
+		StringMatches(``),
+		HeadersMatch(map[string][]string{
+			"Exit-Status-Code": {"1"},
+		}),
+		HeaderKeysNotIncluded([]string{"Content-Type"}),
+	)
+}